Code reorg, cleanup and way better errors from queries
