+++ relayer/query.go
@@ -34,6 +34,10 @@ import (
 
 // TODO: Validate all info coming back from these queries using the verifier
 
+//////////////////////////////
+//    ICS 02 -> CLIENTS     //
+//////////////////////////////
+
 // QueryConsensusState returns a consensus state for a given chain to be used as a
 // client in another chain, fetches latest height when passed 0 as arg
 func (c *Chain) QueryConsensusState(height int64) (*tmclient.ConsensusState, error) {
@@ -42,6 +46,7 @@ func (c *Chain) QueryConsensusState(height int64) (*tmclient.ConsensusState, err
 		validators *ctypes.ResultValidators
 		err        error
 	)
+
 	if height == 0 {
 		commit, err = c.Client.Commit(nil)
 		validators, err = c.Client.Validators(nil, 1, 10000)
@@ -49,8 +54,9 @@ func (c *Chain) QueryConsensusState(height int64) (*tmclient.ConsensusState, err
 		commit, err = c.Client.Commit(&height)
 		validators, err = c.Client.Validators(nil, 1, 10000)
 	}
+
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("query cons state failed: %w", err)
 	}
 
 	state := &tmclient.ConsensusState{
@@ -79,16 +85,21 @@ func (c *Chain) QueryClientConsensusState(srcHeight, dstHeight int64) (clientTyp
 
 	res, err := c.QueryABCI(req)
 	if err != nil {
-		return conStateRes, err
+		return conStateRes, qClntConsStateErr(err)
+	} else if res.Value == nil {
+		// TODO: Better way to handle this?
+		return clientTypes.NewConsensusStateResponse("notfound", nil, nil, 0), nil
 	}
 
 	var cs exported.ConsensusState
 	if err := c.Amino.UnmarshalBinaryLengthPrefixed(res.Value, &cs); err != nil {
-		return conStateRes, err
+		return conStateRes, qClntConsStateErr(err)
 	}
 
 	return clientTypes.NewConsensusStateResponse(c.PathEnd.ClientID, cs, res.Proof, res.Height), nil
 }
+
+func qClntConsStateErr(err error) error { return fmt.Errorf("query client cons state failed: %w", err) }
 
 // QueryClientState retrevies the latest consensus state for a client in state at a given height
 func (c *Chain) QueryClientState() (clientTypes.StateResponse, error) {
@@ -105,38 +116,50 @@ func (c *Chain) QueryClientState() (clientTypes.StateResponse, error) {
 
 	res, err := c.QueryABCI(req)
 	if err != nil {
-		return conStateRes, err
+		return conStateRes, qClntStateErr(err)
+	} else if res.Value == nil {
+		// TODO: Better way to handle this?
+		return clientTypes.NewClientStateResponse("notfound", nil, nil, 0), nil
 	}
 
 	var cs exported.ClientState
 	if err := c.Amino.UnmarshalBinaryLengthPrefixed(res.Value, &cs); err != nil {
-		return conStateRes, err
+		return conStateRes, qClntStateErr(err)
 	}
 
 	return clientTypes.NewClientStateResponse(c.PathEnd.ClientID, cs, res.Proof, res.Height), nil
 }
 
+func qClntStateErr(err error) error { return fmt.Errorf("query client state failed: %w", err) }
+
 // QueryClients queries all the clients!
 func (c *Chain) QueryClients(page, limit int) ([]clientExported.ClientState, error) {
-	params := clientTypes.NewQueryAllClientsParams(page, limit)
-	bz, err := c.Cdc.MarshalJSON(params)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal query params: %w", err)
+	var (
+		bz      []byte
+		err     error
+		clients []clientExported.ClientState
+	)
+
+	if bz, err = c.Cdc.MarshalJSON(clientTypes.NewQueryAllClientsParams(page, limit)); err != nil {
+		return nil, qClntsErr(err)
 	}
 
-	route := fmt.Sprintf("custom/%s/%s/%s", ibctypes.QuerierRoute, clientTypes.QuerierRoute, clientTypes.QueryAllClients)
-	res, _, err := c.QueryWithData(route, bz)
-	if err != nil {
-		return nil, err
+	if bz, _, err = c.QueryWithData(ibcQuerierRoute(clientTypes.QuerierRoute, clientTypes.QueryAllClients), bz); err != nil {
+		return nil, qClntsErr(err)
 	}
 
-	var clients []clientExported.ClientState
-	err = c.Cdc.UnmarshalJSON(res, &clients)
-	if err != nil {
-		return nil, fmt.Errorf("failed to unmarshal light clients: %w", err)
+	if err = c.Cdc.UnmarshalJSON(bz, &clients); err != nil {
+		return nil, qClntsErr(err)
 	}
+
 	return clients, nil
 }
+
+func qClntsErr(err error) error { return fmt.Errorf("query clients failed: %w", err) }
+
+//////////////////////////////
+//  ICS 03 -> CONNECTIONS   //
+//////////////////////////////
 
 // QueryConnectionsUsingClient gets any connections that exist between chain and counterparty
 func (c *Chain) QueryConnectionsUsingClient(height int64) (clientConns connTypes.ClientConnectionsResponse, err error) {
@@ -153,15 +176,19 @@ func (c *Chain) QueryConnectionsUsingClient(height int64) (clientConns connTypes
 
 	res, err := c.QueryABCI(req)
 	if err != nil {
-		return clientConns, err
+		return clientConns, qConnsUsingClntsErr(err)
 	}
 
 	var paths []string
 	if err := c.Amino.UnmarshalBinaryLengthPrefixed(res.Value, &paths); err != nil {
-		return clientConns, err
+		return clientConns, qConnsUsingClntsErr(err)
 	}
 
 	return connTypes.NewClientConnectionsResponse(c.PathEnd.ClientID, paths, res.Proof, res.Height), nil
+}
+
+func qConnsUsingClntsErr(err error) error {
+	return fmt.Errorf("query connections using clients failed: %w", err)
 }
 
 // QueryConnection returns the remote end of a given connection
@@ -179,28 +206,32 @@ func (c *Chain) QueryConnection(height int64) (connTypes.ConnectionResponse, err
 
 	res, err := c.QueryABCI(req)
 	if err != nil {
-		return connTypes.ConnectionResponse{}, err
+		return connTypes.ConnectionResponse{}, qConnErr(err)
 	} else if res.Value == nil {
-		return connTypes.ConnectionResponse{Connection: connTypes.ConnectionEnd{State: connState.UNINITIALIZED}}, nil
+		// NOTE: This is returned so that the switch statement in ConnectionStep works properly
+		return emptyConnRes, nil
 	}
 
 	var connection connTypes.ConnectionEnd
 	if err := c.Amino.UnmarshalBinaryLengthPrefixed(res.Value, &connection); err != nil {
-		return connTypes.ConnectionResponse{}, err
+		return connTypes.ConnectionResponse{}, qConnErr(err)
 	}
 
 	return connTypes.NewConnectionResponse(c.PathEnd.ConnectionID, connection, res.Proof, res.Height), nil
 }
 
-// QueryChannelsUsingConnections returns all channels associated with a given set of connections
-func (c *Chain) QueryChannelsUsingConnections(connections []string) ([]chanTypes.ChannelResponse, error) {
-	return []chanTypes.ChannelResponse{}, nil
-}
+func qConnErr(err error) error { return fmt.Errorf("query connection failed: %w", err) }
+
+var emptyConnRes = connTypes.ConnectionResponse{Connection: connTypes.ConnectionEnd{State: connState.UNINITIALIZED}}
+
+//////////////////////////////
+//    ICS 04 -> CHANNEL     //
+//////////////////////////////
 
 // QueryChannel returns the channel associated with a channelID
-func (c *Chain) QueryChannel(height int64) (chanTypes.ChannelResponse, error) {
+func (c *Chain) QueryChannel(height int64) (chanRes chanTypes.ChannelResponse, err error) {
 	if err := c.PathEnd.Validate(CHANPATH); !c.PathSet() && err != nil {
-		return chanTypes.ChannelResponse{}, c.ErrPathNotSet(CHANPATH, err)
+		return chanRes, c.ErrPathNotSet(CHANPATH, err)
 	}
 
 	req := abci.RequestQuery{
@@ -212,27 +243,53 @@ func (c *Chain) QueryChannel(height int64) (chanTypes.ChannelResponse, error) {
 
 	res, err := c.QueryABCI(req)
 	if err != nil {
-		return chanTypes.ChannelResponse{}, err
+		return chanRes, qChanErr(err)
 	} else if res.Value == nil {
-		return chanTypes.ChannelResponse{Channel: chanTypes.Channel{State: chanState.UNINITIALIZED}}, nil
+		// NOTE: This is returned so that the switch statement in ChannelStep works properly
+		return emptyChanRes, nil
 	}
 
 	var channel chanTypes.Channel
 	if err := c.Amino.UnmarshalBinaryLengthPrefixed(res.Value, &channel); err != nil {
-		return chanTypes.ChannelResponse{}, err
+		return chanRes, qChanErr(err)
 	}
 
 	return chanTypes.NewChannelResponse(c.PathEnd.PortID, c.PathEnd.ChannelID, channel, res.Proof, res.Height), nil
 }
 
-// func (c *Chain) QueryChannels(height int64) ([]chanTypes.ChannelResponse, error) {
-// 	return
-// }
+func qChanErr(err error) error { return fmt.Errorf("query channel failed: %w", err) }
+
+var emptyChanRes = chanTypes.ChannelResponse{Channel: chanTypes.Channel{State: chanState.UNINITIALIZED}}
+
+// QueryChannels returns all the channels that are registered on a chain
+func (c *Chain) QueryChannels(page, limit int) ([]chanTypes.Channel, error) {
+	var (
+		bz       []byte
+		err      error
+		channels []chanTypes.Channel
+	)
+
+	if bz, err = c.Cdc.MarshalJSON(chanTypes.NewQueryAllChannelsParams(page, limit)); err != nil {
+		return nil, qChansErr(err)
+	}
+
+	if bz, _, err = c.QueryWithData(ibcQuerierRoute(chanTypes.QuerierRoute, chanTypes.QueryAllChannels), bz); err != nil {
+		return nil, qChansErr(err)
+	}
+
+	if err = c.Cdc.UnmarshalJSON(bz, &channels); err != nil {
+		return nil, qChansErr(err)
+	}
+
+	return channels, nil
+}
+
+func qChansErr(err error) error { return fmt.Errorf("query channels failed: %w", err) }
 
 // QueryNextSeqRecv returns the next seqRecv for a configured channel
-func (c *Chain) QueryNextSeqRecv(height int64) (*chanTypes.RecvResponse, error) {
+func (c *Chain) QueryNextSeqRecv(height int64) (recvRes chanTypes.RecvResponse, err error) {
 	if err := c.PathEnd.Validate(CHANPATH); !c.PathSet() && err != nil {
-		return nil, c.ErrPathNotSet(CHANPATH, err)
+		return recvRes, c.ErrPathNotSet(CHANPATH, err)
 	}
 
 	req := abci.RequestQuery{
@@ -244,20 +301,19 @@ func (c *Chain) QueryNextSeqRecv(height int64) (*chanTypes.RecvResponse, error) 
 
 	res, err := c.QueryABCI(req)
 	if err != nil {
-		return nil, err
+		return recvRes, err
 	} else if res.Value == nil {
-		return nil, nil
+		// TODO: figure out how to return not found error
+		return recvRes, nil
 	}
 
-	out := chanTypes.NewRecvResponse(
+	return chanTypes.NewRecvResponse(
 		c.PathEnd.PortID,
 		c.PathEnd.ChannelID,
 		binary.BigEndian.Uint64(res.Value),
 		res.Proof,
 		res.Height,
-	)
-
-	return &out, nil
+	), nil
 }
 
 // QueryNextSeqSend returns the next seqSend for a configured channel
@@ -277,6 +333,7 @@ func (c *Chain) QueryNextSeqSend(height int64) (uint64, error) {
 	if err != nil {
 		return 0, err
 	} else if res.Value == nil {
+		// NOTE: figure out how to return not found error
 		return 0, nil
 	}
 
@@ -284,9 +341,9 @@ func (c *Chain) QueryNextSeqSend(height int64) (uint64, error) {
 }
 
 // QueryPacketCommitment returns the packet commitment proof at a given height
-func (c *Chain) QueryPacketCommitment(height, seq int64) (*CommitmentResponse, error) {
+func (c *Chain) QueryPacketCommitment(height, seq int64) (comRes CommitmentResponse, err error) {
 	if err := c.PathEnd.Validate(CHANPATH); !c.PathSet() && err != nil {
-		return nil, c.ErrPathNotSet(CHANPATH, err)
+		return comRes, c.ErrPathNotSet(CHANPATH, err)
 	}
 
 	req := abci.RequestQuery{
@@ -298,12 +355,13 @@ func (c *Chain) QueryPacketCommitment(height, seq int64) (*CommitmentResponse, e
 
 	res, err := c.QueryABCI(req)
 	if err != nil {
-		return nil, err
+		return comRes, qPacketCommitmentErr(err)
 	} else if res.Value == nil {
-		return nil, fmt.Errorf("no response returned")
+		// TODO: Is this the not found error we want to return here?
+		return comRes, nil
 	}
 
-	return &CommitmentResponse{
+	return CommitmentResponse{
 		Data:  res.Value,
 		Proof: commitmenttypes.MerkleProof{Proof: res.Proof},
 		ProofPath: commitmenttypes.NewMerklePath(
@@ -316,6 +374,10 @@ func (c *Chain) QueryPacketCommitment(height, seq int64) (*CommitmentResponse, e
 	}, nil
 }
 
+func qPacketCommitmentErr(err error) error {
+	return fmt.Errorf("query packet commitment failed: %w", err)
+}
+
 // CommitmentResponse returns the commiment hash along with the proof data
 // NOTE: CommitmentResponse is used to wrap query response from querying PacketCommitment AND PacketAcknowledgement
 type CommitmentResponse struct {
@@ -326,9 +388,9 @@ type CommitmentResponse struct {
 }
 
 // QueryPacketAck returns the packet commitment proof at a given height
-func (c *Chain) QueryPacketAck(height, seq int64) (*CommitmentResponse, error) {
+func (c *Chain) QueryPacketAck(height, seq int64) (comRes CommitmentResponse, err error) {
 	if err := c.PathEnd.Validate(CHANPATH); !c.PathSet() && err != nil {
-		return nil, c.ErrPathNotSet(CHANPATH, err)
+		return comRes, c.ErrPathNotSet(CHANPATH, err)
 	}
 
 	req := abci.RequestQuery{
@@ -340,12 +402,12 @@ func (c *Chain) QueryPacketAck(height, seq int64) (*CommitmentResponse, error) {
 
 	res, err := c.QueryABCI(req)
 	if err != nil {
-		return nil, err
+		return comRes, qPacketAckErr(err)
 	} else if res.Value == nil {
-		return nil, fmt.Errorf("no response returned")
+		return comRes, nil
 	}
 
-	return &CommitmentResponse{
+	return CommitmentResponse{
 		Data:  res.Value,
 		Proof: commitmenttypes.MerkleProof{Proof: res.Proof},
 		ProofPath: commitmenttypes.NewMerklePath(
@@ -358,42 +420,43 @@ func (c *Chain) QueryPacketAck(height, seq int64) (*CommitmentResponse, error) {
 	}, nil
 }
 
+func qPacketAckErr(err error) error {
+	return fmt.Errorf("query packet acknowledgement failed: %w", err)
+}
+
 // QueryTxs returns an array of transactions given a tag
-func (c *Chain) QueryTxs(height uint64, events []string) (*sdk.SearchTxsResult, error) {
+func (c *Chain) QueryTxs(height uint64, events []string) (res sdk.SearchTxsResult, err error) {
 	if len(events) == 0 {
-		return nil, errors.New("must declare at least one event to search")
+		return res, errors.New("must declare at least one event to search")
 	}
 
 	resTxs, err := c.Client.TxSearch(strings.Join(events, " AND "), true, 0, 10000, "asc")
 	if err != nil {
-		return nil, err
+		return res, err
 	}
 
 	for _, tx := range resTxs.Txs {
-		err := c.ValidateTxResult(tx)
-		if err != nil {
-			return nil, err
+		if err = c.ValidateTxResult(tx); err != nil {
+			return res, err
 		}
 	}
 
 	resBlocks, err := c.queryBlocksForTxResults(resTxs.Txs)
 	if err != nil {
-		return nil, err
+		return res, err
 	}
 
 	txs, err := c.formatTxResults(resTxs.Txs, resBlocks)
 	if err != nil {
-		return nil, err
+		return res, err
 	}
 
-	result := sdk.NewSearchTxsResult(resTxs.TotalCount, len(txs), 0, 10000, txs)
-
-	return &result, nil
+	return sdk.NewSearchTxsResult(resTxs.TotalCount, len(txs), 0, 10000, txs), nil
 }
 
 // QueryABCI is an affordance for querying the ABCI server associated with a chain
 // Similar to cliCtx.QueryABCI
-func (c *Chain) QueryABCI(req abci.RequestQuery) (abci.ResponseQuery, error) {
+func (c *Chain) QueryABCI(req abci.RequestQuery) (res abci.ResponseQuery, err error) {
 	opts := rpcclient.ABCIQueryOptions{
 		Height: req.GetHeight(),
 		Prove:  req.Prove,
@@ -401,11 +464,11 @@ func (c *Chain) QueryABCI(req abci.RequestQuery) (abci.ResponseQuery, error) {
 
 	result, err := c.Client.ABCIQueryWithOptions(req.Path, req.Data, opts)
 	if err != nil {
-		return abci.ResponseQuery{}, err
+		return res, err
 	}
 
 	if !result.Response.IsOK() {
-		return abci.ResponseQuery{}, errors.New(result.Response.Log)
+		return res, errors.New(result.Response.Log)
 	}
 
 	// data from trusted node or subspace query doesn't need verification
@@ -414,26 +477,20 @@ func (c *Chain) QueryABCI(req abci.RequestQuery) (abci.ResponseQuery, error) {
 	}
 
 	if err = c.VerifyProof(req.Path, result.Response); err != nil {
-		return abci.ResponseQuery{}, err
+		return res, err
 	}
 
 	return result.Response, nil
 }
 
 // QueryWithData satisfies auth.NodeQuerier interface and used for fetching account details
-func (c *Chain) QueryWithData(path string, data []byte) ([]byte, int64, error) {
-	req := abci.RequestQuery{
-		Path:   path,
-		Height: 0,
-		Data:   data,
+func (c *Chain) QueryWithData(p string, d []byte) (byt []byte, i int64, err error) {
+	var res abci.ResponseQuery
+	if res, err = c.QueryABCI(abci.RequestQuery{Path: p, Height: 0, Data: d}); err != nil {
+		return byt, i, err
 	}
 
-	resp, err := c.QueryABCI(req)
-	if err != nil {
-		return []byte{}, 0, err
-	}
-
-	return resp.Value, resp.Height, nil
+	return res.Value, res.Height, nil
 }
 
 // QueryLatestHeight queries the chain for the latest height and returns it
@@ -441,9 +498,7 @@ func (c *Chain) QueryLatestHeight() (int64, error) {
 	res, err := c.Client.Status()
 	if err != nil {
 		return -1, err
-	}
-
-	if res.SyncInfo.CatchingUp {
+	} else if res.SyncInfo.CatchingUp {
 		return -1, fmt.Errorf("node at %s running chain %s not caught up", c.RPCAddr, c.ChainID)
 	}
 
@@ -462,10 +517,6 @@ func (h *heights) err() error {
 		out = fmt.Errorf("err: %w ", err)
 	}
 	return out
-}
-
-func (h *heights) out() map[string]int64 {
-	return h.Map
 }
 
 // QueryLatestHeights returns the heights of multiple chains at once
@@ -489,21 +540,18 @@ func QueryLatestHeights(chains ...*Chain) (map[string]int64, error) {
 		}(hs, &wg, chain)
 	}
 	wg.Wait()
-	return hs.out(), hs.err()
+	return hs.Map, hs.err()
 }
 
 // QueryLatestHeader returns the latest header from the chain
-func (c *Chain) QueryLatestHeader() (*tmclient.Header, error) {
-	h, err := c.QueryLatestHeight()
-	if err != nil {
+func (c *Chain) QueryLatestHeader() (out *tmclient.Header, err error) {
+	var h int64
+	if h, err = c.QueryLatestHeight(); err != nil {
 		return nil, err
 	}
-
-	out, err := c.QueryHeaderAtHeight(h)
-	if err != nil {
+	if out, err = c.QueryHeaderAtHeight(h); err != nil {
 		return nil, err
 	}
-
 	return out, nil
 }
 
@@ -554,18 +602,15 @@ func isQueryStoreWithProof(path string) bool {
 // queryBlocksForTxResults returns a map[blockHeight]txResult
 func (c *Chain) queryBlocksForTxResults(resTxs []*ctypes.ResultTx) (map[int64]*ctypes.ResultBlock, error) {
 	resBlocks := make(map[int64]*ctypes.ResultBlock)
-
 	for _, resTx := range resTxs {
 		if _, ok := resBlocks[resTx.Height]; !ok {
 			resBlock, err := c.Client.Block(&resTx.Height)
 			if err != nil {
 				return nil, err
 			}
-
 			resBlocks[resTx.Height] = resBlock
 		}
 	}
-
 	return resBlocks, nil
 }
 
@@ -579,7 +624,6 @@ func (c *Chain) formatTxResults(resTxs []*ctypes.ResultTx, resBlocks map[int64]*
 			return nil, err
 		}
 	}
-
 	return out, nil
 }
 
@@ -589,7 +633,6 @@ func (c *Chain) formatTxResult(resTx *ctypes.ResultTx, resBlock *ctypes.ResultBl
 	if err != nil {
 		return sdk.TxResponse{}, err
 	}
-
 	return sdk.NewResponseResultTx(resTx, tx, resBlock.Block.Time.Format(time.RFC3339)), nil
 }
 
@@ -603,4 +646,8 @@ func parseTx(cdc *codec.Codec, txBytes []byte) (sdk.Tx, error) {
 	}
 
 	return tx, nil
+}
+
+func ibcQuerierRoute(module, path string) string {
+	return fmt.Sprintf("custom/%s/%s/%s", ibctypes.QuerierRoute, module, path)
 }
